+++ wagtail/contrib/forms/views.py
@@ -9,7 +9,7 @@ from django.utils.translation import ngettext
 from django.views.generic import ListView, TemplateView
 
 from wagtail.admin import messages
-from wagtail.admin.views.reports import SpreadsheetExportMixin
+from wagtail.admin.views.mixins import SpreadsheetExportMixin
 from wagtail.contrib.forms.forms import SelectDateForm
 from wagtail.contrib.forms.utils import get_forms_for_user
 from wagtail.core.models import Page
